(PUP-1843) Refactor duplicate code
Puppet::Indirector::Hiera and Puppet::DataBinding::Hiera had
identical implementations.  Puppet::DataBinding::Hiera is now a subclass
of Puppet::Indirector::Hiera.  Common behavior in their specs was also
moved into spec/shared_behaviors.
