+++ lib/puppet/indirector/data_binding/hiera.rb
@@ -1,50 +1,7 @@
-require 'puppet/indirector/code'
+require 'puppet/indirector/hiera'
 require 'hiera/scope'
 
-class Puppet::DataBinding::Hiera < Puppet::Indirector::Code
+class Puppet::DataBinding::Hiera < Puppet::Indirector::Hiera
   desc "Retrieve data using Hiera."
-
-  def initialize(*args)
-    if ! Puppet.features.hiera?
-      raise "Hiera terminus not supported without hiera library"
-    end
-    super
-  end
-
-  if defined?(::Psych::SyntaxError)
-    DataBindingExceptions = [::StandardError, ::Psych::SyntaxError]
-  else
-    DataBindingExceptions = [::StandardError]
-  end
-
-  def find(request)
-    hiera.lookup(request.key, nil, Hiera::Scope.new(request.options[:variables]), nil, nil)
-  rescue *DataBindingExceptions => detail
-    raise Puppet::DataBinding::LookupError.new(detail.message, detail)
-  end
-
-  private
-
-  def self.hiera_config
-    hiera_config = Puppet.settings[:hiera_config]
-    config = {}
-
-    if Puppet::FileSystem.exist?(hiera_config)
-      config = Hiera::Config.load(hiera_config)
-    else
-      Puppet.warning "Config file #{hiera_config} not found, using Hiera defaults"
-    end
-
-    config[:logger] = 'puppet'
-    config
-  end
-
-  def self.hiera
-    @hiera ||= Hiera.new(:config => hiera_config)
-  end
-
-  def hiera
-    self.class.hiera
-  end
 end
